refactor(fs.scandir): reorganize code
