+++ packages/fs/fs.scandir/src/providers/sync.spec.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import * as sinon from 'sinon';
 import { Dirent, DirentType, Stats, StatsMode } from '@nodelib/fs.macchiato';
 
-import Settings from '../settings';
+import { Settings } from '../settings';
 import * as utils from '../utils';
 import * as provider from './sync';
