+++ packages/fs/fs.scandir/src/providers/sync.ts
@@ -3,7 +3,7 @@ import * as fsStat from '@nodelib/fs.stat';
 import * as utils from '../utils';
 import * as common from './common';
 
-import type Settings from '../settings';
+import type { Settings } from '../settings';
 import type { Entry, ErrnoException } from '../types';
 
 export function read(directory: string, settings: Settings): Entry[] {
