+++ packages/fs/fs.scandir/src/settings.spec.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import { Stats } from '@nodelib/fs.macchiato';
 
 import * as fs from './adapters/fs';
-import Settings from './settings';
+import { Settings } from './settings';
 
 describe('Settings', () => {
 	it('should return instance with default values', () => {
