+++ packages/fs/fs.scandir/src/types/index.ts
@@ -9,6 +9,7 @@ export interface Entry {
 
 export type Stats = fs.Stats;
 export type ErrnoException = NodeJS.ErrnoException;
+export type AsyncCallback = (error: ErrnoException | null, entries: Entry[]) => void;
 
 export interface Dirent {
 	isBlockDevice: () => boolean;
